Restructure benchmark logic for better readability
