+++ wrapper/src/api/providers/benchmark/cache/cache.ts
@@ -1,0 +1,7 @@
+import { InvalidatableBenchmarkCache } from './invalidatableCache';
+import { BenchmarkCacheContent } from './types';
+
+export abstract class BenchmarkCache<
+  Config,
+  Content extends BenchmarkCacheContent<Config>
+> extends InvalidatableBenchmarkCache<Config, Content> {}
