+++ wrapper/src/api/providers/benchmark/cache/invalidatableCache.ts
@@ -1,0 +1,149 @@
+import { tables } from '../../../database';
+import {
+  DatasetId,
+  ExperimentConfigItem,
+  ExperimentId,
+  SimilarityThresholdFunctionId,
+} from '../../../server/types';
+import { BasicBenchmarkCache } from './basicCache';
+import { BenchmarkCacheBaseConfig, BenchmarkCacheContent } from './types';
+
+export abstract class InvalidatableBenchmarkCache<
+  Config,
+  Content extends BenchmarkCacheContent<Config>
+> extends BasicBenchmarkCache<Config, Content> {
+  static invalidateExperiment(experiment: ExperimentId): void {
+    this.invalidate((instance) =>
+      instance.invalidateKeys(instance.experimentKeys, experiment)
+    );
+  }
+
+  static invalidateDataset(dataset: DatasetId): void {
+    this.invalidate((instance) =>
+      instance.invalidateKeys(instance.datasetKeys, dataset)
+    );
+  }
+
+  static invalidateSimilarityFunction(
+    func: SimilarityThresholdFunctionId
+  ): void {
+    this.invalidate((instance) =>
+      instance.invalidateKeys(instance.similarityFunctionKeys, func)
+    );
+  }
+
+  protected static instances: WeakRef<
+    InvalidatableBenchmarkCache<unknown, BenchmarkCacheContent<unknown>>
+  >[] = [];
+
+  protected static invalidate(
+    run: (
+      instance: InvalidatableBenchmarkCache<
+        unknown,
+        BenchmarkCacheContent<unknown>
+      >
+    ) => void
+  ): void {
+    this.instances = this.instances.filter((instance) => instance.deref());
+    for (const instance of this.instances.map(
+      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+      (instance) => instance.deref()!
+    )) {
+      run(instance);
+    }
+  }
+
+  protected static invalidatePersistentCaches(identifier: string): void {
+    tables.cache.intersectionCounts.delete(
+      {
+        key: `%${identifier}%`,
+      },
+      'LIKE'
+    );
+  }
+
+  protected datasetKeys = new Map<DatasetId, Set<string>>();
+  protected similarityFunctionKeys = new Map<
+    SimilarityThresholdFunctionId,
+    Set<string>
+  >();
+  protected experimentKeys = new Map<ExperimentId, Set<string>>();
+
+  constructor() {
+    super();
+    InvalidatableBenchmarkCache.instances.push(new WeakRef(this));
+  }
+
+  protected invalidateKeys(keys: Map<number, Set<string>>, id: number): void {
+    [...this.getInvalidationKeys(keys, id).values()].map((key) =>
+      this.cache.del(key)
+    );
+    keys.delete(id);
+  }
+
+  protected getInvalidationKeys(
+    keys: Map<number, Set<string>>,
+    id: number
+  ): Set<string> {
+    let keySet = keys.get(id);
+    if (!keySet) {
+      keySet = new Set();
+      keys.set(id, keySet);
+    }
+    return keySet;
+  }
+
+  protected createAndCache(
+    config: BenchmarkCacheBaseConfig,
+    key: string
+  ): Content {
+    this.prepareInvalidation(key, config);
+    return super.createAndCache(config, key);
+  }
+
+  protected prepareInvalidation(
+    key: string,
+    { datasetId, group2: negative, group1: positive }: BenchmarkCacheBaseConfig
+  ): void {
+    this.getInvalidationKeys(this.datasetKeys, datasetId).add(key);
+    this.prepareExperimentAndSimilarityInvalidation(key, positive);
+    this.prepareExperimentAndSimilarityInvalidation(key, negative);
+  }
+
+  protected prepareExperimentAndSimilarityInvalidation(
+    key: string,
+    items: ExperimentConfigItem[]
+  ): void {
+    for (const { experimentId, similarity } of items) {
+      this.getInvalidationKeys(this.experimentKeys, experimentId).add(key);
+      if (similarity) {
+        this.getInvalidationKeys(
+          this.similarityFunctionKeys,
+          similarity.func
+        ).add(key);
+      }
+    }
+  }
+
+  protected dispose(key: string, content: Content): void {
+    const { datasetId, group1, group2 } = this.mapCustomConfigToBaseConfig(
+      content.config
+    );
+    this.datasetKeys.get(datasetId)?.delete(key);
+    this.deleteExperimentConfigItemsKey(group1, key);
+    this.deleteExperimentConfigItemsKey(group2, key);
+    super.dispose(key, content);
+  }
+
+  protected deleteExperimentConfigItemsKey(
+    items: ExperimentConfigItem[],
+    key: string
+  ): void {
+    for (const { experimentId, similarity } of items) {
+      this.experimentKeys.get(experimentId)?.delete(key);
+      if (similarity) {
+        this.similarityFunctionKeys.get(similarity.func)?.delete(key);
+      }
+    }
+  }
+}
