+++ wrapper/src/api/providers/benchmark/intersection/base.ts
@@ -1,19 +1,16 @@
 import { tables } from '../../../database';
 import { LazyProperty } from '../../../tools/lazyProperty';
 import { StoredDataset } from '../../dataset/util/converter';
+import { BenchmarkCacheContent } from '../cache';
 import { Clustering } from '../cluster/types';
 import { UnionFind } from '../cluster/unionFind';
+import { SubclusterCache } from '../subclusteringCache/subclusteringCache';
 import type { Intersection as IntersectionSubclass } from '.';
-import {
-  IntersectionCache,
-  IntersectionCacheContent,
-  IntersectionConfig,
-  SubclusterCache,
-} from './cache';
+import { IntersectionCache, IntersectionConfig } from './intersectionCache';
 import { IntersectionQueries } from './queries';
 
 export class IntersectionBase
-  implements IntersectionCacheContent<IntersectionConfig> {
+  implements BenchmarkCacheContent<IntersectionConfig> {
   get clustering(): Clustering {
     return this._clustering.value;
   }
