--- wrapper/src/api/providers/benchmark/intersection/cache/index.ts
@@ -1,3 +1,0 @@
-export * from './base';
-export * from './intersectionCache';
-export * from './subclusteringCache';
