+++ wrapper/src/api/providers/benchmark/intersection/index.ts
@@ -1,4 +1,3 @@
 import { IntersectionClusters } from './clusters';
-export { IntersectionCache } from './cache';
 
 export class Intersection extends IntersectionClusters {}
